Enable Text to be used by Viewer outside Files app
Since Nextcloud 18.0.4 the Files app triggers the "LoadViewer" event, so
that event can be used to register Text with Viewer instead of doing it
through the "OCA\Files::loadAdditionalScripts" event (although the
previous way needs to be kept for now just in case Text is run on a
previous Nextcloud 18 version).

Besides being more correct for the Files app it also makes possible to
use Text when opening files through Viewer in other apps, like Talk.

Note, however, that in public share pages Text is still opened using the
fallback actions instead of through Viewer, as Viewer currently does not
provide the share token of the public shared folder that contains the
file to be opened.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>